+++ NEWS.md
@@ -2,4 +2,4 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * Improved documentation 
-* Fixes to `OEIS_links`
+* Fixes to `OEIS_links`, `OEIS_mathematica`